Reorganize module files into subfolders and renaming
* Moved subnet_computation.py into configure subfolder
* Renamed pcluster.py into commands.py
* Moved cfnconfig.py into config subfolder
* Moved config_sanitiy.py into config subfolder
