+++ cli/pcluster/config/config_sanity.py
