+++ cli/tests/pcluster/configure/test_subnet_cidr.py
@@ -1,5 +1,5 @@
 from assertpy import assert_that
-from pcluster.subnet_computation import evaluate_cidr, get_subnet_cidr
+from pcluster.configure.subnet_computation import evaluate_cidr, get_subnet_cidr
 
 
 def test_empty_vpc():
